Ignore duplicate members when generating types.
This commit fixes #26 by ensuring that duplicate members are ignored
completely when generating types. That is, if a member is inherited twice
via different interfaces, the generated mock will not include an
implementation for that member. Doing so would necessitate either generating
only a single member, or making one or both implementations explicit. Either
of these choices could go against the expectations of the user, so it is
safest to not generate anything in that case, thus forcing the user to provide
the appropriate implementation.