Refactor IPlugin interface
In order to accommodate recursive application of plugins, the first step
is to refactor the IPlugin interface such that less responsibility (and
flexibility) is put onto the plugin implementations. Instead of being
responsible for generating all the code for configuring behavior, plugins
are now queried for a default value for each symbol in the mocked
interface.